Cast explicitly.
git-svn-id: https://svn.dealii.org/trunk@22111 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/multigrid/mg_renumbered_03.cc
@@ -183,7 +183,7 @@ void OutputCreator<dim>::cell(
   const FEValuesBase<dim>& fe = info.fe_values();
   const std::vector<std::vector<double> >& uh = info.values[0];
 
-  const unsigned int square_root = std::pow(fe.n_quadrature_points, 1./dim)+.5;
+  const unsigned int square_root = static_cast<unsigned int>(std::pow(fe.n_quadrature_points, 1./dim)+.5);
   for (unsigned int k1=0; k1<square_root; ++k1)
   {
     for (unsigned int k2=0; k2<square_root; ++k2)
